Исправил ошибку в отображении профиля клиента
