+++ src/types/clientTypes.ts
@@ -21,3 +21,25 @@ export interface IClient {
   trainingLevel: string;
   physicalData: string;
 }
+
+export interface IClientProfile {
+  _id: string;
+  user: {
+    _id: string;
+    email: string;
+    role: string;
+    firstName: string;
+    lastName: string;
+    phoneNumber: string;
+    gender: string;
+    dateOfBirth: string;
+    timeZone: {
+      value: string;
+      offset: string;
+    };
+  };
+  subscribes: string[];
+  preferredWorkoutType: string;
+  trainingLevel: string;
+  physicalData: string;
+}
